Fixed one type in a class and enhanced recent games output Also switched to a relative assembly path instead of a public one

+++ LibOfLegends/com/riotgames/platform/statistics/PlayerGameStats.cs
@@ -46,7 +46,7 @@ namespace com.riotgames.platform.statistics
 		public string subType;
 		public string queueType;
 		public bool premadeTeam;
-		public float predictedWinPct;
+		public double predictedWinPct;
 		public int rating;
 		public int championId;
 	}
